Migrate ParticipantCountsOVerTime Job to use Cloud Scheduler instead of CRON
